requirements.txt
pandas>=2.1.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
streamlit>=1.28.0

# LangGraph / LangChain stack
langgraph
langchain
langchain-openai

# FastAPI / API server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

